+++ config/network/config.go
@@ -6,17 +6,19 @@ import (
 	"time"
 
 	"github.com/spf13/viper"
+
+	"github.com/onflow/flow-go/network/p2p/config"
 )
 
 // Config encapsulation of configuration structs for all components related to the Flow network.
 type Config struct {
 	// UnicastRateLimitersConfig configuration for all unicast rate limiters.
-	UnicastRateLimitersConfig `mapstructure:",squash"`
-	ResourceManagerConfig     `mapstructure:",squash"`
-	ConnectionManagerConfig   `mapstructure:",squash"`
+	UnicastRateLimitersConfig    `mapstructure:",squash"`
+	config.ResourceManagerConfig `mapstructure:",squash"`
+	ConnectionManagerConfig      `mapstructure:",squash"`
 	// GossipSubConfig core gossipsub configuration.
-	GossipSubConfig `mapstructure:",squash"`
-	AlspConfig      `mapstructure:",squash"`
+	config.GossipSubConfig `mapstructure:",squash"`
+	AlspConfig             `mapstructure:",squash"`
 
 	// NetworkConnectionPruning determines whether connections to nodes
 	// that are not part of protocol state should be trimmed
