+++ network/p2p/inspector/validation/inspect_message_request.go
@@ -6,7 +6,7 @@ import (
 	pubsub_pb "github.com/libp2p/go-libp2p-pubsub/pb"
 	"github.com/libp2p/go-libp2p/core/peer"
 
-	netconf "github.com/onflow/flow-go/config/network"
+	netconf "github.com/onflow/flow-go/network/p2p/config"
 	"github.com/onflow/flow-go/network/p2p/inspector/internal"
 )
